pyproject.toml
[project]
name = "coursework5"
version = "0.1.0"
description = ""
authors = [
    {name = "Sepega",email = "twnue2020@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-telegram-bot (>=22.4,<23.0)",
    "pytelegrambotapi (>=4.29.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
